requirements.txt
greenlet==3.3.1
psycopg2-binary==2.9.11
PyMySQL==1.1.2
pyodbc==5.3.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
